requirements.txt
base58==2.0.0
cryptography==2.8
ecdsa==0.15
mnemonic==0.19
PyJWT==1.7.1
requests==2.23.0
retrying==1.3.3
web3~=5.12.1